% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission-utils.R
\name{add_friendly_names}
\alias{add_friendly_names}
\title{Add user-friendly submission name}
\usage{
add_friendly_names(data)
}
\arguments{
\item{data}{The submission data in the form given by
[synapseforms::make_tidier_table].}
}
\value{
Original data with a submission column
  in the second column index, with submission names.
}
\description{
Add a user-friendly submission name column in the
form of the submitter's last name, the
compound, and the formDataId to ensure uniqueness.
Setup as "lastname - compound (formDataId)".
}
