% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gather-clean.R
\name{add_section_variables}
\alias{add_section_variables}
\title{Add unanswered questions within a section.}
\usage{
add_section_variables(data, lookup_table)
}
\arguments{
\item{data}{Dataframe with columns "section", "variable", and "exp_num".}

\item{lookup_table}{Dataframe with columns "section",
"step", "variable" , and "label" used for user-friendly section and
variable display. "step" maps desired "section" names. "label" maps
desired "variable" names.}
}
\description{
Uses `lookup_table` data to add in questions within a section that were
unanswered (and therefore missing from the original JSON data).
}
