% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{append_clinical_to_submission}
\alias{append_clinical_to_submission}
\title{Append clinical multiplier to submissions}
\usage{
append_clinical_to_submission(submissions)
}
\arguments{
\item{submissions}{Data containing one or more submissiosn}
}
\value{
Data from `submissions` with an added `clinical` column containing
  the multiplier
}
\description{
Given a data frome of one or more submissions, finds the clinical
}
