% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gather-clean.R
\name{append_exp_nums}
\alias{append_exp_nums}
\title{Append experiment numbers to step name}
\usage{
append_exp_nums(data)
}
\arguments{
\item{data}{Data frame containing submission data}
}
\description{
When data has multiple experiments, appends the experiment number to each
section, e.g. `LD50 [1]`, `LD50 [2]`, etc.
}
