% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_denominator}
\alias{calculate_denominator}
\title{Calculate denominator}
\usage{
calculate_denominator(data)
}
\arguments{
\item{data}{Data frame containing the submission data}
}
\value{
A number representing the denominator by which the score should be
  divided
}
\description{
Calculates the denominator by which the score should be divided
}
