% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_scores_rowwise}
\alias{calculate_scores_rowwise}
\title{Calculate scores rowwise}
\usage{
calculate_scores_rowwise(reviews, submissions)
}
\arguments{
\item{reviews}{Dataframe review table.}

\item{submissions}{Data frame of submissions *including* clinical multiplier
(i.e. the output from [append_clinical_to_submission()]).}
}
\description{
Given a data frame of scores, adds a column with the weighted score
calculated.
}
