% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_section_score}
\alias{calculate_section_score}
\title{Calculate the score for a section}
\usage{
calculate_section_score(data, lookup, score = 1, species = 1, clinical = 1)
}
\arguments{
\item{data}{Data frame containing the section data. Should have columns
`section`, `variable`, and `response` at a minimum.}

\item{lookup}{Lookup table that contains the partial beta weights for
applicable variables.}

\item{score}{Average score given by reviewers.}

\item{species}{Within-species (usually 0.67) or across-species (usually 0.33)
multiplication factor. Not used for sections `basic`, `pk_in_silico` or
`pk_in_vivo`.}

\item{clinical}{Clinical (usually 0.67) or preclinical (usually 0.33)
multiplication factor. Not used for sections `naming`, `basic`,
`pk_in_silico` or `pk_in_vivo`.}
}
\value{
A numeric value indicating the score for the section
}
\description{
Given data for a single section of a submission, calculate that section's
score based on multiplication of the clinical vs. preclinical factor, within-
vs. across-species factor, section factor, partial beta weight, and score
given by the reviewer (or average score across multiple reviewers).
}
