% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_submission_score}
\alias{calculate_submission_score}
\title{Score an entire submission}
\usage{
calculate_submission_score(data, reviews, lookup)
}
\arguments{
\item{data}{Data frame containining the submission. Should have columns
`section`, `step`, `variable`, and `response` at a minimum.}

\item{reviews}{Data frame of scores submitted by reviewers.}

\item{lookup}{Lookup table that contains the partial beta weights for
applicable variables.}
}
\value{
The score for the submission
}
\description{
Given a data frame containing a submission and a table of reviews, calculate
the score for a submission. First, the score for each section is calculated
by multiplying the reviewer's score by the appropriate multipliers (see
[calculate_section_score()]). If multiple reviewers scored the section, the
section score is calculated for each and averaged.
}
\details{
Then, the scores across all sections are added and divided by the denominator
(see [calculate_denominator()]).
}
\seealso{
[calculate_section_score()]

[calculate_denominator()]
}
