% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_experiment_variables}
\alias{clean_experiment_variables}
\title{Clean experiment variables}
\usage{
clean_experiment_variables(data)
}
\arguments{
\item{data}{The submission data in the form given by
[synapseforms::make_tidier_table].}
}
\description{
Multiple experiments lead to `sub_variable`s with names of the
form "age_range1", "age_range2". This will remove the number at
the end of `sub_variable` and append to the `section` name.
Note that this only happens if the number is > 0. This is due
to having one known `sub_variable`, ID50, that ends in a number,
but is not related to the multiple experiment issue.
}
