% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gather-clean.R
\name{combine_route_responses}
\alias{combine_route_responses}
\title{Combine multiple routes into one comma-separated response}
\usage{
combine_route_responses(data)
}
\arguments{
\item{data}{List containing route data}
}
\description{
Combine multiple routes into one comma-separated response
}
