% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gather-clean.R
\name{create_section_table}
\alias{create_section_table}
\title{Create table for a section}
\usage{
create_section_table(data, section)
}
\arguments{
\item{data}{A list containing data from one section of a submission}

\item{section}{The section name}
}
\description{
Create table for a section of a submission. Some sections contain multiple
experiments nested within them; if this is the case, they will be unnested
and given an experiment number.
}
