% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gather-clean.R
\name{create_table_by_sections}
\alias{create_table_by_sections}
\title{Create table for a section}
\usage{
create_table_by_sections(data, section)
}
\arguments{
\item{data}{A list containing data from one section of a submission}

\item{section}{The section name}
}
\description{
Create table for a section. Nested lists in the data will
be unnested into separate rows.
}
