% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gather-clean.R
\name{create_table_from_values}
\alias{create_table_from_values}
\title{Create a tibble from the values in section}
\usage{
create_table_from_values(data, section, exp_num = NA)
}
\arguments{
\item{data}{A list containing data from one section of a submission}

\item{section}{The section name}

\item{exp_num}{Numeric experiment number}
}
\description{
Create a tibble with section name, experiment number, variables, and
response values.
}
