% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_submissions_table}
\alias{format_submissions_table}
\title{Format submissions table}
\usage{
format_submissions_table(data)
}
\arguments{
\item{data}{Submissions table as given by
[synapseforms::download_all_and_get_table()]}
}
\description{
Format the submissions table to be
tidier, have genrated submission names,
and cleaned up variables.
}
