% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{geom_mean_score}
\alias{geom_mean_score}
\title{Calculate geometric mean of non-zero scores}
\usage{
geom_mean_score(values)
}
\arguments{
\item{values}{Values to average}
}
\description{
Discards any zero or NA scores, then calculates the geometric mean of the
remaining scores. If all scores are zero, returns 0.
}
