% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-review-section.R
\name{get_submission_list}
\alias{get_submission_list}
\title{Get list of submissions with ids}
\usage{
get_submission_list(data)
}
\arguments{
\item{data}{The submission data as given by
[get_submissions()].}
}
\value{
named list where names are the submission names
  and values are their form_data_ids.
}
\description{
Get list of submissions with their form_data_id.
}
