% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gather-clean.R
\name{make_clean_table}
\alias{make_clean_table}
\title{Make clean submissions table}
\usage{
make_clean_table(data, section_lookup_table, variable_lookup_table)
}
\arguments{
\item{data}{The submission data in the form given by
[synapseforms::make_tidier_table], plus columns step and section.}

\item{section_lookup_table}{Dataframe with columns "section" and
"step" used for user-friendly section names}

\item{variable_lookup_table}{Dataframe with columns "variable" and
"label" used for user-friendly variable names}
}
\description{
Make submissions table clean by fixing experiment
numbering and adding in user-friendly step and label
columns instead of json variable styled section and
variable columns.
}
\keyword{internal}
