% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gather-clean.R
\name{map_names}
\alias{map_names}
\title{Append user-friendly section and variable names}
\usage{
map_names(data, lookup_table, complete = TRUE)
}
\arguments{
\item{data}{Dataframe with columns "section", "variable", and "exp_num".}

\item{lookup_table}{Dataframe with columns "section",
"step", "variable" , and "label" used for user-friendly section and
variable display. "step" maps desired "section" names. "label" maps
desired "variable" names.}

\item{complete}{If `TRUE`, will join in all section and variable names that
were not provided as part of the submission. If `FALSE`, will only return
the data that was present in the JSON file.}
}
\description{
Appends columns "step" and "label", which corresponds with "section" and
"variable". Map via lookup_table and fix missing step/label.
}
