% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-review_section.R
\name{mod_review_section_ui}
\alias{mod_review_section_ui}
\alias{mod_review_section_server}
\title{mod_review_section_ui and mod_review_section_server}
\usage{
mod_review_section_ui(id)

mod_review_section_server(input, output, session, data)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
Review and score the selected section of a submission
}
\keyword{internal}
