% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gather-clean.R
\name{remove_empty_objects}
\alias{remove_empty_objects}
\title{ALZ-157: Remove empty objects from inner lists for legacy submissions.}
\usage{
remove_empty_objects(data_list)
}
\arguments{
\item{data}{List containing data}
}
\description{
ALZ-157: Remove empty objects from inner lists for legacy submissions.
}
