% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gather-clean.R
\name{therapeutic_approach_response}
\alias{therapeutic_approach_response}
\title{Rename response "both" to "prophylactic, symptomatic" in therapeutic approach}
\usage{
therapeutic_approach_response(data)
}
\arguments{
\item{data}{Data frame containing submission data}
}
\description{
Rename response "both" to "prophylactic, symptomatic" in therapeutic approach
}
