% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{transformLegacyPivotLonger}
\alias{transformLegacyPivotLonger}
\title{A legacy pivot_longer function}
\usage{
transformLegacyPivotLonger(clinicals)
}
\arguments{
\item{clinicals}{The list of clinicals}
}
\description{
A legacy pivot_longer function
}
