% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gather-clean.R
\name{add_friendly_names}
\alias{add_friendly_names}
\title{Add user-friendly submission name}
\usage{
add_friendly_names(data)
}
\arguments{
\item{data}{The submission data in the form given by
[synapseforms::make_tidier_table], plus columns step and section.}
}
\value{
List of submission names
}
\description{
Add a user-friendly submission name column in the
form of the submitter's last name - the
compound. Note that these are not guaranteed to be
unique and submissions should be referred to by their
form_data_id.
}
