% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_section_score}
\alias{calculate_section_score}
\title{Calculate the score for a section}
\usage{
calculate_section_score(
  data,
  lookup,
  section = 1,
  clinical = 1,
  species = 1,
  score = 1
)
}
\arguments{
\item{data}{Data frame containing the section data. Should have columns
`section`, `variable`, and `response` at a minimum.}

\item{lookup}{Lookup table that contains the partial beta weights for
applicable variables.}

\item{section}{Section multiplication factor. Usually 1 except for sections
`pk_in_silico`, `pk_in_vitro`, and `pk_in_vivo`.}

\item{clinical}{Clinical (usually 0.67) or preclinical (usually 0.33)
multiplication factor.}

\item{species}{Within-species (usually 0.67) or across-species (usually 0.33)
multiplication factor.}

\item{score}{Average score given by reviewers.}
}
\value{
A numeric value indicating the score for the section
}
\description{
Given data for a single section of a submission, calculate that section's
score based on multiplication of the clinical vs. preclinical factor, within-
vs. across-species factor, section factor, partial beta weight, and score
given by the reviewer (or average score across multiple reviewers).
}
