% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gather-clean.R
\name{change_logical_responses}
\alias{change_logical_responses}
\title{Change logical responses to yes/no}
\usage{
change_logical_responses(data)
}
\arguments{
\item{data}{The submission data in the form given by
[synapseforms::make_tidier_table], plus columns step and section.}
}
\description{
Change TRUE/FALSE responses to be yes/no.
}
