% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gather-clean.R
\name{clean_experiment_variables}
\alias{clean_experiment_variables}
\title{Clean experiment variables}
\usage{
clean_experiment_variables(data)
}
\arguments{
\item{data}{The submission data in the form given by
[synapseforms::make_tidier_table], plus columns step and section.}
}
\description{
Multiple experiments lead to `variable`s with names of the
form "age_range1", "age_range2". This will remove the number at
the end of `variable` and append to the `step` name.
Note that this only happens if the number is > 0. This is due
to having one known `variable`, ID50, that ends in a number,
but is not related to the multiple experiment issue.
}
