% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gather-clean.R
\name{create_table_by_sections}
\alias{create_table_by_sections}
\title{Create table by sections}
\usage{
create_table_by_sections(data, section, unnest = FALSE)
}
\arguments{
\item{data}{Submission list from JSON}

\item{section}{The section name}

\item{unnest}{TRUE to unnest the responses; FALSE to have the responses
nested in a list}
}
\description{
Create table by going through each section.
}
