% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gather-clean.R
\name{create_table_from_json_file}
\alias{create_table_from_json_file}
\title{Create table from JSON file}
\usage{
create_table_from_json_file(filename, data_id)
}
\arguments{
\item{filename}{Path to JSON file}

\item{data_id}{Data file handle ID}
}
\description{
Convert a JSON file containing submission data to a data frame.
}
