% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gather-clean.R
\name{create_table_from_values}
\alias{create_table_from_values}
\title{Create a tibble from the values in section}
\usage{
create_table_from_values(data, section, exp_num = NA, unnest)
}
\arguments{
\item{data}{Submission list from JSON}

\item{section}{The section name}

\item{exp_num}{Numeric experiment number}

\item{unnest}{TRUE to unnest the responses; FALSE to have the responses
nested in a list}
}
\description{
Create a tibble with section name, experiment number, variables, and
response values.
}
