% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gather-clean.R
\name{create_values_table}
\alias{create_values_table}
\title{Create a tibble from the values within section}
\usage{
create_values_table(data, section, lookup_table, complete = TRUE, exp_num = NA)
}
\arguments{
\item{data}{A list containing data from one section of a submission}

\item{section}{The section name}

\item{lookup_table}{Dataframe with columns "section",
"step", "variable" , and "label" used for user-friendly section and
variable display. "step" maps desired "section" names. "label" maps
desired "variable" names.}

\item{complete}{If `TRUE`, will join in all section and variable names that
were not provided as part of the submission. If `FALSE`, will only return
the data that was present in the JSON file.}

\item{exp_num}{Numeric experiment number}
}
\description{
Create a tibble with section name, experiment number, variables, and
response values.
}
