% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-review-section.R
\name{get_sections}
\alias{get_sections}
\title{Get submission sections}
\usage{
get_sections(data, submission_name)
}
\arguments{
\item{data}{The submission data as given by
[get_submissions()].}

\item{submission_name}{Submission name.}
}
\description{
Get submission steps, which can be thought of
as the main sections.
}
