% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-review-section.R
\name{get_sections}
\alias{get_sections}
\title{Get submission section names}
\usage{
get_sections(data, submission)
}
\arguments{
\item{data}{All submissions data.}

\item{submission}{Name of the submission.}
}
\description{
Get submission section names.
}
\keyword{internal}
