% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gather-clean.R
\name{get_submissions}
\alias{get_submissions}
\title{Get the submissions based on status}
\usage{
get_submissions(
  syn,
  group,
  statuses,
  section_lookup_table,
  variable_lookup_table
)
}
\arguments{
\item{syn}{Synapse client object (e.g. output of `synapse$Synapse()`)}

<<<<<<< HEAD
=======
\item{group}{The number for a specific Synapse forms group.}

>>>>>>> master
\item{statuses}{A character vector of statuses to
include from the set: "Accepted", "Rejected",
"In Review".}

<<<<<<< HEAD
\item{section_lookup_table}{Dataframe with columns "variable" and
=======
\item{section_lookup_table}{Dataframe with columns "section" and
"step" used for user-friendly section names}

\item{variable_lookup_table}{Dataframe with columns "variable" and
>>>>>>> master
"label" used for user-friendly variable names}
}
\description{
Get the submissions based on status.
}
