% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gather-clean.R
\name{get_submissions}
\alias{get_submissions}
\title{Get the submissions based on status}
\usage{
get_submissions(syn, group, statuses, lookup_table)
}
\arguments{
\item{syn}{Synapse client object (e.g. output of `synapse$Synapse()`)}

\item{group}{The number for a specific Synapse forms group.}

\item{statuses}{A character vector of statuses to
include from the set: `SUBMITTED_WAITING_FOR_REVIEW`,
`ACCEPTED`, `REJECTED`.}

\item{lookup_table}{Dataframe with columns "section",
"step", "variable" , and "label" used for user-friendly section and
variable display. "step" maps desired "section" names. "label" maps
desired "variable" names.}
}
\description{
Get the submissions based on status.
}
