% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-view-all-section.R
\name{get_submissions}
\alias{get_submissions}
\title{Get the submissions based on status}
\usage{
get_submissions(syn, group, statuses)
}
\arguments{
\item{statuses}{A character vector of statuses to
include from the set: "Accepted", "Rejected",
"In Review".}
}
\description{
Get the submissions based on status.
}
