% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-panel-section.R
\name{mod_panel_section_ui}
\alias{mod_panel_section_ui}
\alias{mod_panel_section_server}
\title{mod_panel_section_ui and mod_panel_section_server}
\usage{
mod_panel_section_ui(id)

mod_panel_section_server(
  input,
  output,
  session,
  synapse,
  syn,
  user,
  submissions,
  reviews_table,
  submissions_table,
  partial_betas
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{synapse}{Synapse client (e.g. output of
`reticulate::import("synapseclient")`)}

\item{syn}{Synapse client object (e.g. output of `synapse$Synapse()`)}

\item{submissions}{Data frame containing tidied submissions}

\item{reviews_table}{Synapse table that stores the scores and comments}

\item{submissions_table}{Synapse table that stores the
overall submission scores and comments}
}
\description{
View summarized scores for a submission
}
\keyword{internal}
