% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-review-section.R
\name{mod_review_section_ui}
\alias{mod_review_section_ui}
\alias{mod_review_section_server}
\title{mod_review_section_ui and mod_review_section_server}
\usage{
mod_review_section_ui(id)

mod_review_section_server(
  input,
  output,
  session,
  synapse,
  syn,
  reviews_table,
  section_lookup_table,
  variable_lookup_table
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{synapse}{Synapse client (e.g. output of
`reticulate::import("synapseclient")`)}

\item{syn}{Synapse client object (e.g. output of `synapse$Synapse()`)}

\item{reviews_table}{Synapse table that stores the scores and comments}

<<<<<<< HEAD
\item{section_lookup_table}{Dataframe with columns "variable" and
=======
\item{section_lookup_table}{Dataframe with columns "section" and
"step" used for user-friendly section names}

\item{variable_lookup_table}{Dataframe with columns "variable" and
>>>>>>> master
"label" used for user-friendly variable names}
}
\description{
Review and score the selected section of a submission
}
\keyword{internal}
