% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-review-section.R
\name{mod_review_section_ui}
\alias{mod_review_section_ui}
\alias{mod_review_section_server}
\title{mod_review_section_ui and mod_review_section_server}
\usage{
mod_review_section_ui(id)

mod_review_section_server(
  input,
  output,
  session,
  synapse,
  syn,
  user,
  submissions,
  reviews_table
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{synapse}{Synapse client (e.g. output of
`reticulate::import("synapseclient")`)}

\item{syn}{Synapse client object (e.g. output of `synapse$Synapse()`)}

\item{submissions}{Data frame containing tidied submissions}

\item{reviews_table}{Synapse table that stores the scores and comments}
}
\description{
Review and score the selected section of a submission
}
\keyword{internal}
