% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-view-all-section.R
\name{mod_view_all_section_ui}
\alias{mod_view_all_section_ui}
\alias{mod_view_all_section_server}
\title{mod_view_all_section_ui and mod_view_all_section_server}
\usage{
mod_view_all_section_ui(id)

mod_view_all_section_server(
  input,
  output,
  session,
  synapse,
  syn,
  group,
  lookup_table,
  sub_metadata
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{synapse}{Synapse client (e.g. output of
`reticulate::import("synapseclient")`)}

\item{syn}{Synapse client object (e.g. output of `synapse$Synapse()`)}
}
\description{
View and search among all submissions
}
\keyword{internal}
