% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-panel-section.R
\name{pull_reviews_table}
\alias{pull_reviews_table}
\title{Pull latest review table}
\usage{
pull_reviews_table(syn, reviews_table)
}
\arguments{
\item{syn}{Synapse client object (e.g. output of `synapse$Synapse()`)}

\item{reviews_table}{Synapse table that stores the scores and comments}
}
\description{
Pull latest review table from Synapse.
}
\keyword{internal}
