% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-panel-section.R
\name{show_review_table}
\alias{show_review_table}
\title{Show review table}
\usage{
show_review_table(input, output, reviews, submission_id)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{reviews}{Dataframe review table.}

\item{submission_id}{Reactive shiny object containing submission id
accessible via `submission()`.}
}
\description{
Show review table.
}
\keyword{internal}
