% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{transformLegacyPivotLonger}
\alias{transformLegacyPivotLonger}
\title{Score an entire submission}
\usage{
transformLegacyPivotLonger(clinicals)
}
\arguments{
\item{submission}{Data frame containining the submission. Should have columns
`section`, `step`, `variable`, and `response` at a minimum.}

\item{reviews}{Data frame containing weighted scores (output of
[calculate_scores_rowwise()], or [pull_reviews_table()]).}
}
\value{
The score for the submission
}
\description{
Given a data frame containing a submission and a table of reviews, calculate
the score for a submission. First, the score for each section is calculated
by multiplying the reviewer's score by the appropriate multipliers (see
[calculate_section_score()]). If multiple reviewers scored the section, the
section score is calculated for each and averaged.
}
\details{
Then, the scores across all sections are added and divided by the denominator
(see [calculate_denominator()]).
}
\seealso{
[calculate_scores_rowwise()]

[pull_reviews_table()]

[calculate_denominator()]
}
